bin/config/requirements.txt
#django==3.0.7
uwsgi==2.0.17.1
psycopg2-binary==2.7.6.1
flake8==3.2.1
